Dedupe components by adding generics
