+++ src/components/lineChart/hooks/useSewerWaterChartOptions.ts
@@ -1,12 +1,30 @@
 import { SeriesLineOptions, SeriesScatterOptions } from 'highcharts';
 import { useMemo } from 'react';
-import { TimeframeOption, getFilteredValues } from '~/utils/timeframe';
-import { RegionalSewerPerInstallationValue } from '~/types/data';
 import { formatDateFromSeconds } from '~/utils/formatDate';
 import { formatNumber } from '~/utils/formatNumber';
 import { getItemFromArray } from '~/utils/getItemFromArray';
 import { replaceVariablesInText } from '~/utils/replaceVariablesInText';
-import { Value, TranslationStrings } from '../regionalSewerWaterChart';
+import { getFilteredValues, TimeframeOption } from '~/utils/timeframe';
+
+export interface SewerPerInstallationBaseValue {
+  date_measurement_unix: number;
+  rna_normalized: number;
+  rwzi_awzi_name: string;
+}
+
+export type Value = {
+  date: number;
+  value?: number;
+  week_start_unix: number;
+  week_end_unix: number;
+};
+
+export type TranslationStrings = {
+  average_label_text: string;
+  secondary_label_text: string;
+  range_description: string;
+  daily_label_text: string;
+};
 
 type Week = {
   start: number;
@@ -44,9 +62,11 @@ function createRemainingDaysData(value: Value | undefined, maxDate: number) {
   });
 }
 
-export function useRegionalSewerWaterChartOptions(
+export function useSewerWaterChartOptions<
+  T extends SewerPerInstallationBaseValue
+>(
   averageValues: Value[],
-  scatterPlotValues: RegionalSewerPerInstallationValue[],
+  scatterPlotValues: T[] | undefined,
   text: TranslationStrings,
   timeframe: TimeframeOption,
   selectedRWZI?: string
@@ -59,11 +79,13 @@ export function useRegionalSewerWaterChartOptions(
 
   const filteredScatterPlotValues = useMemo(
     () =>
-      getFilteredValues(
-        scatterPlotValues,
-        timeframe,
-        (value) => value.date_measurement_unix * 1000
-      ),
+      scatterPlotValues
+        ? getFilteredValues(
+            scatterPlotValues,
+            timeframe,
+            (value) => value.date_measurement_unix * 1000
+          )
+        : [],
     [scatterPlotValues, timeframe]
   );
