+++ src/components/lineChart/sewer-water-chart.tsx
@@ -5,34 +5,26 @@ import styled from 'styled-components';
 import Dot from '~/assets/dot.svg';
 import Line from '~/assets/line.svg';
 import { ValueAnnotation } from '~/components-styled/value-annotation';
-import { RegionalSewerPerInstallationValue } from '~/types/data';
 import { TimeframeOption } from '~/utils/timeframe';
-import { useRegionalSewerWaterChartOptions } from './hooks/useRegionalSewerWaterChartOptions';
+import {
+  SewerPerInstallationBaseValue,
+  TranslationStrings,
+  useSewerWaterChartOptions,
+  Value,
+} from './hooks/useSewerWaterChartOptions';
 
-export type Value = {
-  date: number;
-  value?: number;
-  week_start_unix: number;
-  week_end_unix: number;
-};
-
-export type TProps = {
+export type TProps<T extends SewerPerInstallationBaseValue> = {
   averageValues: Value[];
-  scatterPlotValues: RegionalSewerPerInstallationValue[];
+  scatterPlotValues?: T[];
   text: TranslationStrings;
   timeframe: TimeframeOption;
   selectedInstallation?: string;
   valueAnnotation?: string;
 };
 
-export type TranslationStrings = {
-  average_label_text: string;
-  secondary_label_text: string;
-  range_description: string;
-  daily_label_text: string;
-};
-
-export function RegionalSewerWaterChart(props: TProps) {
+export function SewerWaterChart<T extends SewerPerInstallationBaseValue>(
+  props: TProps<T>
+) {
   const {
     averageValues,
     scatterPlotValues,
@@ -42,7 +34,7 @@ export function RegionalSewerWaterChart(props: TProps) {
     valueAnnotation,
   } = props;
 
-  const chartOptions = useRegionalSewerWaterChartOptions(
+  const chartOptions = useSewerWaterChartOptions(
     averageValues,
     scatterPlotValues,
     text,
@@ -56,21 +48,23 @@ export function RegionalSewerWaterChart(props: TProps) {
         <ValueAnnotation mb={2}>{valueAnnotation}</ValueAnnotation>
       )}
       <HighchartsReact highcharts={Highcharts} options={chartOptions} />
-      <div>
-        <ul>
-          {chartOptions.series
-            ?.filter((serie) => serie.description?.length)
-            .map((serie) => (
-              <LegendItem key={serie.name}>
-                <LegendMarker>
-                  {serie.type === 'scatter' && <Dot fill={serie.color} />}
-                  {serie.type === 'line' && <Line stroke={serie.color} />}
-                </LegendMarker>
-                <div>{serie.description}</div>
-              </LegendItem>
-            ))}
-        </ul>
-      </div>
+      {scatterPlotValues && (
+        <div>
+          <ul>
+            {chartOptions.series
+              ?.filter((serie) => serie.description?.length)
+              .map((serie) => (
+                <LegendItem key={serie.name}>
+                  <LegendMarker>
+                    {serie.type === 'scatter' && <Dot fill={serie.color} />}
+                    {serie.type === 'line' && <Line stroke={serie.color} />}
+                  </LegendMarker>
+                  <div>{serie.description}</div>
+                </LegendItem>
+              ))}
+          </ul>
+        </div>
+      )}
     </>
   );
 }
